admin can't be edited

+++ app/Http/Controllers/UserController.php
@@ -53,11 +53,15 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        abort_if($user->role === 'ADMIN', 404);
+
         return view('users.edit', compact('user'));
     }
 
     public function update(Request $request, User $user)
     {
+        abort_if($user->role === 'ADMIN', 404);
+
         $credentials = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'username' => ['required', 'string', 'max:255', Rule::unique('users')->ignore($user)],
@@ -75,6 +79,8 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        abort_if($user->role === 'ADMIN', 404);
+
         $user->delete();
 
         return redirect()->route('users.index')
